move cli sql/cassandra tests
