+++ host/cli/sql/utils.go
@@ -1,0 +1,5 @@
+package sql
+
+// NOTE: change this when moving the test files around during refactoring
+// Path to root folder of cadence repo
+const rootRelativePath = "../../../"
